Create permission directly from context, and add context-specific permission
