+++ selene-common/src/main/java/org/dockbox/selene/common/i18n/Permission.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (C) 2020 Guus Lieben
+ *
+ * This framework is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 2.1 of the
+ * License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+ * the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
+ */
+
+package org.dockbox.selene.common.i18n;
+
+import org.dockbox.selene.api.i18n.permissions.AbstractPermission;
+import org.dockbox.selene.api.i18n.permissions.PermissionContext;
+import org.dockbox.selene.api.objects.Exceptional;
+
+public class Permission implements AbstractPermission {
+
+    private final String key;
+    private final PermissionContext context;
+
+    public Permission(String key, PermissionContext context) {
+        this.key = key;
+        this.context = context;
+    }
+
+    public Permission(String key) {
+        this.key = key;
+        this.context = null;
+    }
+
+    @Override
+    public String get() {
+        return this.key;
+    }
+
+    @Override
+    public Exceptional<PermissionContext> getContext() {
+        return Exceptional.ofNullable(this.context);
+    }
+
+    @Override
+    public AbstractPermission withContext(PermissionContext context) {
+        return new Permission(this.key, context);
+    }
+}
